refactor: move reservation-units to spa
