+++ apps/admin-ui/src/component/my-units/UnitReservationsView.tsx
@@ -7,12 +7,12 @@ import { useNavigate, useParams, useSearchParams } from "react-router-dom";
 import { useLocation } from "react-use";
 import styled from "styled-components";
 import { AutoGrid, VerticalFlex } from "@/styles/layout";
-import { OptionType } from "../../common/types";
-import ReservationUnitTypeFilter from "../filters/ReservationUnitTypeFilter";
-import Tags, { getReducer, toTags } from "../lists/Tags";
+import { OptionType } from "@common/types";
+import { ReservationUnitTypeFilter } from "@component/filters";
+import Tags, { getReducer, toTags } from "@component/lists/Tags";
+import { HR } from "@/component/Table";
 import DayNavigation from "./DayNavigation";
 import UnitReservations from "./UnitReservations";
-import { HR } from "@/component/Table";
 
 const HorisontalFlexWrapper = styled.div`
   display: flex;
