+++ apps/admin-ui/src/spa/reservation-units/DataLoader.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { type ApolloError, useQuery } from "@apollo/client";
 import type { Query, QueryReservationUnitsArgs } from "common/types/gql-types";
 import { filterNonNullable } from "common/src/helpers";
-import { SEARCH_RESERVATION_UNITS_QUERY } from "./queries";
-import { FilterArguments } from "./Filters";
-import { useNotification } from "../../context/NotificationContext";
-import Loader from "../Loader";
-import ReservationUnitsTable from "./ReservationUnitsTable";
+import { useNotification } from "@context/NotificationContext";
+import Loader from "@component/Loader";
+import ReservationUnitsTable from "./Table";
 import { More } from "@/component/More";
-import { LARGE_LIST_PAGE_SIZE } from "../../common/const";
-import { combineResults } from "../../common/util";
+import { LARGE_LIST_PAGE_SIZE } from "@common/const";
+import { combineResults } from "@common/util";
+import { SEARCH_RESERVATION_UNITS_QUERY } from "./queries";
+import type { FilterArguments } from "./Filters";
 
 export type Sort = {
   field: string;
